Removed GlobalCandidateCache.
Since we want all communities to manage their own Candidate, the
GlobalCandidateCache is no longer needed.

Several Candidate related methods have been moved from dispersy.py to
community.py, finally, all references to the GlobalCandidateCache have
been rewritten to use candidates from the community directly.
