+++ candidate.py
@@ -139,10 +139,10 @@ class WalkCandidate(Candidate):
         assert is_address(wan_address), wan_address
         return self._lan_address if wan_address[0] == self._wan_address[0] else self._wan_address
 
-    def merge(self, dispersy, other):
+    def merge(self, community, other):
         if __debug__:
-            from .dispersy import Dispersy
-        assert isinstance(dispersy, Dispersy), type(dispersy)
+            from .community import Community
+        assert isinstance(community, Community), type(community)
         assert isinstance(other, WalkCandidate), type(other)
         self._associations.update(other._associations)
 
@@ -153,7 +153,6 @@ class WalkCandidate(Candidate):
                 self._timestamps[cid] = timestamps
 
                 # TODO: this should be improved
-                community = dispersy._communities.get(cid, None)
                 community.add_candidate(self)
 
         for cid, global_time in self._global_times.iteritems():
